requirements.txt
aiogram==2.19
aiohttp==3.8.1
aiosignal==1.2.0
APScheduler==3.6.3
astroid==2.7.3
async-timeout==4.0.2
attrs==21.4.0
autopep8==1.5.7
Babel==2.9.1
beautifulsoup4==4.10.0
bs4==0.0.1
cachetools==4.2.2
certifi==2021.10.8
charset-normalizer==2.0.7
click==8.0.3
colorama==0.4.4
distlib==0.3.4
filelock==3.4.2
Flask==2.0.2
frozenlist==1.3.0
futures==2.2.0
googlemaps==4.6.0
goslate==1.5.1
idna==3.3
isort==5.9.3
itsdangerous==2.0.1
Jinja2==3.0.3
lazy-object-proxy==1.6.0
MarkupSafe==2.0.1
mccabe==0.6.1
multidict==6.0.2
numpy==1.21.3
pandas==1.3.4
Pillow==8.4.0
places==0.3.4
platformdirs==2.3.0
pycodestyle==2.7.0
PyDictionary==2.0.1
pyfiglet==0.8.post1
pygame==2.1.2
pylint==2.10.2
pyTelegramBotAPI==4.4.0
python-dateutil==2.8.2
python-telegram-bot==13.11
pytz==2021.3
pytz-deprecation-shim==0.1.0.post0
requests==2.26.0
six==1.16.0
soupsieve==2.2.1
termcolor==1.1.0
toml==0.10.2
tornado==6.1
tzdata==2021.5
tzlocal==4.1
urllib3==1.26.7
virtualenv==20.13.0
Werkzeug==2.0.2
wrapt==1.12.1
yarl==1.7.2
